There are now zero (0) lint errors!
Let's keep it that way!

Along with the name cleanups, it turned out that we weren't using most of
the constants in const.go, so they got deleted instead of renamed.  The
ones we did use got moved to a subpackage so that they could be called
something like geoip.MMapCache or the like.